requirements.txt
Flask
pymongo
flask_pymongo
python-dotenv
